fix: moved functionality from App into StopWatch
